third_party/blink/web_tests/virtual/plz-dedicated-worker/external/wpt/fetch/private-network-access/worker.window-expected.txt

This is a testharness.js-based test.
FAIL treat-as-public to local: failure. assert_equals: response loaded expected false but got true
FAIL treat-as-public to private: failure. assert_equals: response loaded expected false but got true
PASS public to public: success.
Harness: the test ran to completion.
